+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.10.0]
+### Added
+- `sharingService.setAccess(owner, itemId, access)` where access is null | 'org' | 'everyone'
+- `sharingService.shareWithGroup(owner, itemId, groupId, confirmItemControl)`
+- `sharingService.isItemSharedWithGroup(itemId, groupId)`
+### Changed
+- deprecated all previous methods / calls as they were non-deterministic
+
 ## [0.9.0]
 ### Changed
 - `portalService.update` was not working; now it will update the `portalProperties` hash only.